requirements.txt
streamlit
openai
python-dotenv
nltk
